requirements.txt
aenum==2.1.2
aiohttp==3.5.4
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.0
async-timeout==3.0.1
attrs==19.1.0
base58==1.0.3
Brlapi==0.6.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
command-not-found==0.3
coz-bytecode==0.5.1
cryptography==2.7
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
ecdsa==0.13.2
gitdb2==2.0.5
GitPython==2.1.11
httplib2==0.9.2
idna==2.8
idna-ssl==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
logzero==1.5.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mnemonic==0.18
multidict==4.5.2
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
ontology-python-sdk==0.1.11
pbkdf2==1.3
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
punica==0.0.22
punica-neo-boa==0.4.9
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycryptodomex==3.8.2
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
smmap2==2.0.5
system-service==0.3
systemd-python==234
typing-extensions==3.7.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.3
usb-creator==0.3.3
virtualenv==16.4.3
wadllib==1.3.2
websockets==7.0
xkit==0.0.0
yarl==1.3.0
zope.interface==4.3.2
